refactor: trpc folder to organize files for a single router in one folder
