+++ src/lib/trpc/router/assets/index.ts
@@ -1,0 +1,5 @@
+import { createTRPCRouter } from "../../trpc";
+
+// TODO: Implement assets router
+export const assetsRouter = createTRPCRouter({
+})
