+++ src/lib/trpc/router/assets/query.ts
@@ -1,0 +1,6 @@
+// import type { CacheEntry } from "@epic-web/cachified";
+// import { LRUCache } from "lru-cache";
+
+// // const cache = new LRUCache<string, CacheEntry>({
+// //     max: 5,
+// // });
