+++ src/modules/pages/landing.tsx
@@ -6,7 +6,7 @@ import { useAndromedaStore } from "@/zustand/andromeda";
 import React from "react";
 import { ConnectWallet } from "../wallet";
 
-interface Props {}
+interface Props { }
 
 const LandingPage: React.FC<Props> = (props) => {
   const { connectedChain } = useAndromedaStore();
@@ -14,7 +14,7 @@ const LandingPage: React.FC<Props> = (props) => {
   return (
     <GridBackground>
       <div className="flex flex-col min-h-screen items-center justify-center gap-4">
-        <img src="/logo.png" className="w-30" />
+        <img src="/logo.png" className="w-30" alt='logo' />
         <p className="text-4xl font-bold">Andromeda Nextjs Starter Template</p>
         {isLoading ? (
           <p>Loading...</p>
