+++ src/modules/ui/PoweredByLogo.tsx
@@ -12,7 +12,7 @@ const PoweredByLogo: FC<Props> = (props) => {
       rel="noopener noreferrer"
     >
       <div className="fixed left-2 bottom-2 max-w-fit pl-1.5 pr-3 py-1 rounded-lg bg-gray-900 flex items-center gap-2">
-        <img src="/logo.png" className="h-6" />
+        <img src="/logo.png" className="h-6" alt='logo' />
         <span className="text-sm text-white">Powered by Andromeda</span>
       </div>
     </Link>
